Refactor: Transition to Microkernel Architecture
This commit refactors the codebase to adhere to a microkernel architecture by decoupling the core orchestration from specific tool and provider implementations.

- Established 'pkg/protocol' for shared interfaces and DTOs (Dependency Inversion).
- Created 'pkg/kernel' to house the core agent conversation loop and orchestration logic.
- Updated 'cmd/run.go' and 'cmd/gc.go' to act as thin CLI wrappers around the kernel.
- Refactored 'internal/provider' and 'internal/tool' to satisfy the new protocol contracts.
- Improved system stability and extensibility by isolating the core system.
